Minor clean-ups to TinyLfu pull request
