+++ simulator/src/main/java/com/github/benmanes/caffeine/cache/simulator/policy/victim/VictimTuQueuePolicy.java
@@ -15,25 +15,17 @@
  */
 package com.github.benmanes.caffeine.cache.simulator.policy.victim;
 
-import com.github.benmanes.caffeine.cache.simulator.admission.Admittor;
 import com.github.benmanes.caffeine.cache.simulator.admission.TinyLfu;
 import com.github.benmanes.caffeine.cache.simulator.policy.two_queue.TuQueuePolicy;
-import com.github.benmanes.caffeine.cache.simulator.policy.victim.VictimLruPolicy.VictimSettings;
 import com.typesafe.config.Config;
 
 /**
- * An extension to the TuQueue policy to guard the COLD queue by an admission policy.
+ * An extension to the TuQueue policy to guard the WARM and COLD queues by an admission policy.
  *
  * @author ben.manes@gmail.com (Ben Manes)
  */
 public final class VictimTuQueuePolicy extends TuQueuePolicy {
-
   public VictimTuQueuePolicy(String name, Config config) {
-    super(name, config, admittor(config));
-  }
-
-  private static Admittor admittor(Config config) {
-    VictimSettings settings = new VictimSettings(config);
-    return new TinyLfu(settings.admission().eps(), settings.admission().confidence(),settings.admission().sampleSize());
+    super(name, config, new TinyLfu(config));
   }
 }
